Share helpers with unit tests
