+++ dot/rpc/subscription/listeners_integration_test.go
@@ -10,11 +10,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"log"
-	"net/http"
-	"net/http/httptest"
 	"os"
-	"strings"
 	"testing"
 	"time"
 
@@ -28,7 +24,6 @@ import (
 	"github.com/ChainSafe/gossamer/lib/runtime/wasmer"
 	"github.com/ChainSafe/gossamer/lib/transaction"
 	"github.com/ChainSafe/gossamer/pkg/scale"
-	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
@@ -289,42 +284,6 @@ func TestGrandpaJustification_Listen(t *testing.T) {
 		require.NoError(t, sub.Stop())
 		wsconn.Wsconn.Close()
 	})
-}
-
-func setupWSConn(t *testing.T) (*WSConn, *websocket.Conn, func()) {
-	t.Helper()
-
-	wskt := new(WSConn)
-	var up = websocket.Upgrader{
-		CheckOrigin: func(r *http.Request) bool { return true },
-	}
-
-	h := func(w http.ResponseWriter, r *http.Request) {
-		c, err := up.Upgrade(w, r, nil)
-		if err != nil {
-			log.Print("error while setup handler:", err)
-			return
-		}
-
-		wskt.Wsconn = c
-	}
-
-	server := httptest.NewServer(http.HandlerFunc(h))
-	defer server.Close()
-
-	wsURL := "ws" + strings.TrimPrefix(server.URL, "http")
-	ws, r, err := websocket.DefaultDialer.Dial(wsURL, nil)
-	defer r.Body.Close()
-
-	require.NoError(t, err)
-
-	cancel := func() {
-		server.Close()
-		ws.Close()
-		wskt.Wsconn.Close()
-	}
-
-	return wskt, ws, cancel
 }
 
 func TestRuntimeChannelListener_Listen(t *testing.T) {
